requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-environ==0.11.2
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0